Make homedir functions work on Windows
